Move jazz functions to new module
