+++ src/MusicManipulations.jl
@@ -7,5 +7,7 @@ using Reexport
 include("midifiles.jl")
 include("quantize.jl")
 include("statistics.jl")
+include("jazz.jl")
+export Jazz
 
 end
